Clean package drop share and republish
This check in will trigger a dry run, in which files are not actually
deleted. Instead the path of the packages to remove will be log. This is
to ensure the deleting logic is safe.

Once the change is verified a follow up change will be made to ensure
files are actually deleted.